requirements.txt
androguard==3.3.5
tensorflow-gpu==1.14.0
scikit-learn==0.20.3
numpy==1.15.4
pyyaml==3.12